# Parse rule identifiers and references from report XML