--- a/app/models/rule_identifier.rb
+++ b/app/models/rule_identifier.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+# Stores the identifier for a rule
+class RuleIdentifier < ApplicationRecord
+  belongs_to :rule
+end