--- a/app/models/rule_identifier.rb
+++ b/app/models/rule_identifier.rb
@@ -3,4 +3,7 @@
 # Stores the identifier for a rule
 class RuleIdentifier < ApplicationRecord
   belongs_to :rule
+
+  validates :label, presence: true
+  validates :system, presence: true
 end