--- a/app/models/rule_reference.rb
+++ b/app/models/rule_reference.rb
@@ -7,6 +7,7 @@ class RuleReference < ApplicationRecord
   has_many :rules, through: :rule_references_rules
 
   validates :label, presence: true, uniqueness: { scope: :href, message: 'and href combination already taken' }
+  validates :href, presence: true, allow_blank: true
 
   def self.from_oscap_objects(oscap_references)
     oscap_references.map do |oscap_reference|