--- a/app/models/rule_references_rule.rb
+++ b/app/models/rule_references_rule.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+# Models rule result references
+class RuleReferencesRule < ApplicationRecord
+  belongs_to :rule_reference
+  belongs_to :rule
+end