--- a/db/migrate/20190517153452_create_rule_references.rb
+++ b/db/migrate/20190517153452_create_rule_references.rb
@@ -1,0 +1,8 @@
+class CreateRuleReferences < ActiveRecord::Migration[5.2]
+  def change
+    create_table :rule_references, id: :uuid do |t|
+      t.string :href
+      t.string :label
+    end
+  end
+end