--- a/db/migrate/20190524170042_create_rule_identifier_table.rb
+++ b/db/migrate/20190524170042_create_rule_identifier_table.rb
@@ -1,0 +1,9 @@
+class CreateRuleIdentifierTable < ActiveRecord::Migration[5.2]
+  def change
+    create_table :rule_identifiers, id: :uuid do |t|
+      t.string :label
+      t.string :system
+      t.references :rule, type: :uuid
+    end
+  end
+end