Towards Aha! COMP-E-99

WIP for:

- More tests


Preliminary performance impact (tested locally with docker-compose):

Before:

```
ParseReportJob JID-1e96bb7b4f72c413dfd91b50 INFO: start
ParseReportJob JID-1e96bb7b4f72c413dfd91b50 INFO: done: 8.186 sec
ParseReportJob JID-72b589705281a3ea232e6fdd INFO: start
ParseReportJob JID-72b589705281a3ea232e6fdd INFO: done: 2.302 sec
ParseReportJob JID-6f71131a8caf4540eee9e718 INFO: start
ParseReportJob JID-6f71131a8caf4540eee9e718 INFO: done: 2.493 sec
ParseReportJob JID-fa8f9b509a16410a4185b4c5 INFO: start
ParseReportJob JID-fa8f9b509a16410a4185b4c5 INFO: done: 2.118 sec
ParseReportJob JID-4a0d0780019476f8974c2e81 INFO: start
ParseReportJob JID-4a0d0780019476f8974c2e81 INFO: done: 2.476 sec
```

After:

```
ParseReportJob JID-1a814121e1c07d9981c1b04f INFO: start
ParseReportJob JID-1a814121e1c07d9981c1b04f INFO: done: 24.544 sec
ParseReportJob JID-9a450dcf2da318efc26c891e INFO: start
ParseReportJob JID-9a450dcf2da318efc26c891e INFO: done: 2.281 sec
ParseReportJob JID-187cbfe45e7b203a1056060a INFO: start
ParseReportJob JID-187cbfe45e7b203a1056060a INFO: done: 2.491 sec
ParseReportJob JID-1aabc589c2fe8cb806e25446 INFO: start
ParseReportJob JID-1aabc589c2fe8cb806e25446 INFO: done: 3.987 sec
```

It appears new imports which create Rule objects will take longer (which is expected), but subsequent imports remain the same as before this change.